{"schema":"decodex/commit/1","summary":"Split autonomy proposal helper modules","authority":"manual"}
